// frontend/app/(authRoutes)/analytics/components/types.ts
/**
 * Copyright 2025 Google LLC
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import { Model } from "@/queries/types";
import { Dispatch, SetStateAction } from "react";

export interface TagsSelectProps {
  readonly selectedTags: string[];
  readonly setSelectedTags: Dispatch<SetStateAction<string[]>>;
}

export interface AverageAnalyticsProps {
  readonly label: string;
  readonly isSecondFilterOn: boolean;
  readonly time: number;
  readonly prompts: number;
  readonly completed: number;
  readonly total: number;
}

export interface FilterType {
  readonly project: string;
  readonly model: Model | null;
  readonly tags: string[];
}